pyproject.toml
[tool.poetry]
name = "gourmet_search"
version = "0.1.0"
description = ""
authors = ["AKI <57143031+aki-033@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
requests = "^2.26.0"
django-environ = "^0.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
